pyproject.toml
[tool.poetry]
name = "tennis-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
pymysql = "^1.1.0"
django-environ = "^0.11.2"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
